Issue #573 Model View Presenter (MVP)
Description: Contains the foundation classes and interfaces for building an
application using an MVP (Model-View-Presenter) design. Also, contains
widget interfaces for the MVP framework that may be implemented in
order to allow views to communicate user-triggered invocations of actions
and changes of state to their presenters in an abstract manner that is
independent of GUI implementation.