Issue #701: Hazard Services Bug Fixes: Responded to comments in patch set 1.  Also, added fixes for the following:
Hazard Information Dialog shows Details (metadata) panel when there is no metadata to view/manipulate for a particular hazard; it should not show the panel.

Hazard Information Dialog sizes its scrollable area to be ridiculously large the first time it is brought up in a H.S. session if the selected hazard event is of a certain hazard type (e.g. FA.W).
Change-Id: I523ed5965a15b8bbfd4ab1d851df0dca2f8f99ad