Issue #701 Changes made to make Simple Hazard Story and Mixed Hazard Story run Hazard Types Table   _overrideLock_   replacedWith SessionManager   Simple and Mixed Hazards -- using replacedBy configuration   Change in Area, Change in Time   Changes to newEventNeeded for Area and Time Change   Change in getContextMenuEntries to show “Remove Potential Events” even   when no events are selected ToolHandler   Fixed bug to test against number of Product Generators, not number of Products
We deleted the groovy unit tests because they are broken and we
agreed we would address testing issues in PV2.

For the modified classpath, we're taking out reference to the
src directory because we don't believe we need it.  However, we
have some doubt about this.  If it turns
out we do, then will also need to add a non-empty src directory
to the master so that Git will pick it up (Git won't keep empty
directories in a REPO).