Issue #1298: Add New Pending to Selected
Added Functionality for Add New Pending to Selected.

Functionality has changed:
In PV1, Add To Selected button would always add to the set of Selected Hazards
In PV2, it is renamed Add New Pending to Selected.
To add to the set of Selected existing hazards, the forecaster uses the Ctl and Shift keys in the Spatial display and Console
When a new hazard is drawn OR new pending hazards are returned by a Recommender e.g. the Dam Break Flood Recommender, then IF the Add New Pending to Selected is ON, they will be added to the set of Selected hazards

Change-Id: Iea381a8a8ca74524dcab8df5b0dd8d1b1661e73d