Issue #2166. Started down the path of creating automated functional tests.  A test for the mixed hazard story is partially complete.
A framework for the automated tests is in place. It should be quite
easy to add additional functional tests.  In fact, a stub simple hazard
story test is already included.

This framework assumes the following.
Suppose object A registers first for event E on the EventBus.
Suppose object B registers for event E on the EventBus.
Then when E occurs, A will be notified before B.
This seems to be holding up in practice so far.  If we have only been
lucky then we will want to figure out a way to allow for the event
bus to have this behavior.

The solution currently has the downside that the code under
test (in HazardServicesAppBuilder) needs to know about the test
classes.  We will want to make this into an extension point to
mitigate this dependency.

Currently, if any test fails then all tests are stopped.  We may
want it to behave more like JUnit whereby the other tests run.

The FunctionalTest class includes some JUnit equivalent methods
like assertFalse.  I tried to bring in the JUnit jar directly via
plugin but then all Heck broke loose.

To run the tests, add
export HAZARD_SERVICES_AUTO_TESTS_ENABLED=true
to eclipse.sh

Then in Hazards, pull down on the View Menu (the arrow at the right
that also lets reset settings/events and select "Run Automated Tests".

You should see the following in your console when it finishes.
DEBUG 2013-10-22 17:31:46,572 [main] CaveLogger: MixedHazardStoryFunctionalTest Successful
DEBUG 2013-10-22 17:31:46,577 [main] CaveLogger: SimpleHazardStoryFunctionalTest Successful
DEBUG 2013-10-22 17:31:46,874 [main] CaveLogger: All tests passed!!!!

You can then immediately try running hazard services normally if you like.
(This is not the case if a test fails for some reason).
Change-Id: I3ca192b8bda2fa428a628b4dc5b352a1f6379a2c