Issue #2336
Changes to allow megawidgets to work well with the product editor,
including:

Addition of multiline option for text field megawidgets.

Configuration option for stateful megawidgets that include GUI elements
that allow the user to rapidly change the state, i.e. several times a
second, to not notify the state change listener until the user completes
the rapid set of state changes.

Addition of a new megawidget, the unbounded list builder, allowing the
user to enter arbitrary strings and to reorder or delete list elements
in order to build up a state consisting of a list of strings.

Creation of a parameters editor factory, which constructs megawidgets
appropriate to parameter types when it is provided with a list of
parameters (Strings, ints, floats, Dates, etc.) and their labels. This
allows megawidgets to be created with very little option specification
required.

Made changes in response to code review comments.

Moved class-keyed hash map class out of ParametersEditorFactory and
promoted it to be a non-inner class, cleaning it up in the process.

Added unit tests for the new class-keyed hash map.

Added manual test of the parameters editor factory and related code;
this consists of a standalone dialog that may be run from the manual
folder in the tests project in order to try out the megawidgets that
are built by this factory.

Added methods to a utility class, and moved classes from the project
gov.noaa.gsd.common.hazards.utilities to gov.noaa.gsd.common.utilities,
since none of the classes that were moved were hazard-specific.

Rebased against latest master changes, and modified tests .classpath
to point to relative-path JAR files.

Change-Id: Ie97af8cab66ea01784022521fbba24d370b0f158