Issue #1462: Hazard Hatching: Responded to comments in Patch Set 1. Rebased against repo.
Fixed merge conflicts.

Changed "Show Hazard Areas" Console View menu option to "Show Hatched Areas".

Made more robust with respect to handling multipolygons and geometry collections.

Cleaned up drawing code to not explicitly check for GeometryCollection.

This changeset allows the forecaster to toggle on/off the display of the actual hazard area
associated with a hazard event. It also changes how hazards are displayed by default.
Hazards are now shown without fill and the geometry border is the color associated with the hazard type.
Fill (or hatching) is reserved to show hazard areas.

Change-Id: I5fb9153dbe524f66877b928958f116c9eec8c959