Issue #2377
Fixed bug that caused time-remaining column's cells to not be
formatted if filtered-out-since-H.S.-startup events are shown
that have active countdowns.

Fixed bug that caused exception when switching Selected Time
Mode from Range to Single.

Fixed bug that caused exception when right-clicking for context
menu over a column header when the horizontal scrollbar is
present and scrolled at least partly to the right.