Issue #2379: Hazard Detail View: Hazard Detail View does not behave correctly when a new hazard category is chosen.
This changeset ensures that if a new hazard category is chosen, then the hazard type
(if one was selected) is reset to an empty string in the HID and only the
default metadata fields are shown.

Also, the hazard type information is removed from the Spatial Display and the Console.

This changeset also addresses how blank fields are represented in the Console. After
discussion with the IWT, it was somewhat clear that sometimes they would like
important fields (such as hazard type) to have a special label if they are undefined
to draw the user's attention to something he/she must do. Instead of adding special
case logic to Hazard Services which tests for specific columns, I made this
configurable as a part of the column definitions in the settings files.

Change-Id: Ie2b056eb0e95c49b9986762fde3d2da7c8c95f9b