Issue #1464: Hazard Clipping and Point Reduction: Added logic to clip hazards to a specified boundary (normally cwa or hsa) and also reduce the number of points in the hazard geometry to a limit specified by configuration in HazardTypes.py
We may need to eventually share the WarnGen polygon point reduction code, but
this cannot not be done at the moment without creating a dependency between WarnGen
and Hazard Services.

For the time being, I'm am trying to use the JTS Douglas Peucker simplification
algorithm to iteratively reduce the number of points in a geometry to the
specified limit.

Change-Id: I1d67cff9415702e3068a01165c1cccc22968518e