Issue #2471: Common interface to message dialogs:
IMessenger is an interface for displaying questions and warnings to the user and
retrieving answers. This allows the viz side (App Builder) to be
responsible for these dialogs, but gives the event manager and other
managers access to them without creating a dependency on the
gov.noaa.gsd.viz.hazards plugin. Since all parts of Hazard Services can
use the same code for creating these dialogs, it makes it easier for them
to be stubbed for testing.