Issue #2471: River Flood Recommender:
Set up the River Flood Recommender to return a point and a polygon instead of two polygons to represent a river forecast point and its associated inundation
area.

This is a directly result of using Geometry Collections. Also, fixed a few
bugs in the drawing code and refined the text labels so that their positions
update dynamically as the user zooms in/out of the display.

Change-Id: I8fd9d02c300f87629de061fa2d75d95603eb8d44