Issue #2471: Context Menu Option to Clip and Reduce Hazards
It was suggested by the IWT that there also be an option on the
right-click context menu to clip and reduce (if necessary)
hazard polygons. So, much like WarnGen, Hazard Services will
clip and reduce hazard polygons on hazard preview and issuance. The
forecaster will also have an option to manually do this via the
context menu.

Clipping and point reduction can always be undone by pressing
the undo button.