Issue #2764: SelectByAreaDbMapResource throws ConcurrentModificationException This is a simple fix, and probably the smallest changeset ever, but I tried several different approaches before deciding on using the CopyOnWriteArrayList.
The problem with the original implementation was that it was possible to have concurrent modification
of the geometry map. This generally occurred while loading large geodatasets from the maps database
such as the ffmp small basins. If the user tried to select ffmp small basins before they were completely
loaded or while a refresh was occurring due to zooming, a ConcurrentModificationException would occur.

CopyOnWriteArrayList objects are costly, but according to javadoc can be useful when transversal operations
vastly outnumber mutations.