Issue #2155: Fixed Editing of Lines and Points:
Added logic to fix a problem with editing lines and moving points.
This was trying to call the storm track recommender, even for lines and
points not associated with a storm track.

Added an enumeration representing possible Hazard Services geometry
types. More refactoring still needs to be done to cleanly implement
this enum, especially in the mouse handlers.

This changeset no longer addresses the problem with the storm track tool.
A cleaner solution needs to be found and this will be easier to do once
Dan's refactor changesets are merged.

Change-Id: I463c6ba35d6fc0ac9b221f5de35acc225b676171