Issue #2155
Fixed a problem that caused a NullPointerException sometimes when
loading a bundle that includes H.S. while H.S. is already running; the
problem was that the old H.S. was going through model-changed events
after it had been disposed of.

Fixed a problem that caused a widget-disposed error occasionally because
the HID had been closed between the scheduling of some asychronous UI
task and the actual running of the task. Also fixed the same sort of
potential problems in the Console.