Issue #2700. Add Product Generator "defineDialog" Functionality to Product Staging Dialog
To really see this changeset in action, you need
to make the change noted in the defineDialog method of
FFA_ProductGenerator.py where you actually return something instead of
an empty Dictionary. Don't forget to run
build_edex subsequently. This change was not made in the
changeset because it breaks all kinds of tests. We
decided we're not quite ready to adjusts the
tests accordingly until Tracy has had a chance to play
around with this capability. So when you test this
changeset, do so first with the defineDialog as is.