Issue #2712: Hazard Information Dialog Tabs: Fixed problem found in testing of patch set 5.
Made changes to HazardServicesMessageHandler to ensure that the new event is modified last
(after the original event) during a "replace-by" operation. This ensures
that the new event is seen as the most recently modified one, and this controls things such as which tab in the HID receives focus when a replace-by action occurs. For example, this ensures that the tab representing a warning replacing a watch gets focus in the HID.
Change-Id: I1235da2baa03e4bbe2fb9caef6e60d2425a9eb61