Issue #2161
Added "Until Further Notice" option for end times of hazards with types
that allow this option (according to HazardTypes.py), and as an option
for "fall below" hazard metadata values for those hazards that have
rise/crest/fall metadata. Option manifests itself as a right-click menu
over rows in the Console, as well as a checkbox next to the End Time or
Fall Below date-time fields in the HID.

When "Until Further Notice" is turned on for a hazard event's end time,
in the Console, appropriate text is shown in the End Time column's cell,
and the end time slider thumb is positioned off the display to the
right. In the HID, the "Until Further Notice" checkbox next to the End
Time date-time fields is checked, said date-time fields are unavailable,
and the end time slider thumb is positioned off the display to the
right.

To support the above, made individual thumbs of multi-value scales
capable of having their editability toggled individually, as opposed to
at widget-level granularity. Also changed time scale megawidget to allow
detail megawidgets to be included next to the date-time fields (to allow
checkbox next to end time), and added new simple CheckBox megawidget
(different from CheckBoxes in that it has a state of true or false, and
only shows a single checkbox with no label above).

Changed Python code to always use new constant UFN_TIME_VALUE, defined
in new VTECConstants.py file, as the end time/fall below value
indicating "Until Further Notice" applies.

Altered use of megawidget side effects applier so that it is applied not
just when a megawidget state change results from user-GUI interaction,
but also when a programmatic setting of a megawidget's state occurs, and
at megawidgets' initialization. This was done in preparation for the
implementation of the fall-below-related until further notice megawidget
interrelationships as a side-effects script; this will be possible when
side effects scripts are supported in the HID (should be in upcoming
code review for Code Improvement #2925). For now, fall-below-related
until further notice functionality is implemented as special-case code
in the HID, marked as @Deprecated.

Changed River Flood Recommender to load its megawidgets side effects
script from a separate file in localization, rather than keeping it
inline as a multi-line string.

Changed HID choice detail time scale megawidgets to be simpler time
megawidgets (without the scale bar), as these take up less space and are
more appropriate when selecting only a single time as opposed to a
range.

In files touched by above, replaced use of Guava collection utilities'
factory method usage with JDK 1.7-compliant constructors of collections,
e.g. new HashSet<>().

Also in these files, corrected Javadoc (using {@link...} instead of
<code>...</code> as appropriate, removing pointless Javadoc for
overridden methods).

Changed until further notice constant names to include "MILLIS" or
"SECS" in order to indicate the time units being used.

Made other, very minor, changes in response to comments.

Change-Id: I5f21ebbb92f5f8e9787f6f40961c6d1983be3bce