Issue #2155
Fixed bug with Settings dialog's visible columns display not staying
synchronized with the state of the console. This was the result of a
lack of defensive copying on the part of the megawidgets.

Fixed bug that caused values set to Until Further Notice to be changed
to another date in the year 2038 if another thumb on the same scale
widget was moved to the right.