Issue #2155
Fixed bugs with date-time fields in time megawidgets that caused
unexpected date-times to be selected when the user manipulated the
drop-down calendar.

Part of the problem was that I had previously attempted to ensure that
setting the date to the current date via the calendar drop-down would
set it to the current CAVE time, circumventing the CDateTime's
hard-coded use of the current system time. However, this code didn't
work correctly, and has thus been removed. Thus, regardless of whether
or not CAVE has its time frozen or otherwise not equivalent to system
time, using the small circular button in the drop-down will always set
the date and time to the current system date and time. This is a small
price to pay; the alternative would be to make copies of the CDateTime
class and many supporting classes, and change the behavior, but that
should only be done if the users deem it necessary.