Issue #3236: Improvements to support Hazard Hatching, Hazard Clipping and Point Reduction.
Undid change which was reading coarse geometry definitions from the
SelectByAreaDbMapResource#getSimpLev method.

Made some efficiency fixes based on Dan's private changeset, "Optimizations
deriving from AFC investigation".  Did not store "isClipped" as a hazard
attribute. Instead, made a it a field in ObservedHazardEvent. Another solution I
thought about was to create a UserGeometry object which contains clipping and
point reduction status and passing this along as user data in a geometry object.
Note that I could not completely eliminate the LegacyProduct_Generator#determineShapeTypesForHazardEvents() method only the geometryList list.
This changeset includes modifications which support:
+ More consistent hatching and clipping in areas with complex geopolitical boundaries and coast lines
+ One label per hazard event. Displaying a label for each polygon associated with a Hazard Event quickly clutters the screen.
+ Better handling of holes in geometries read from the maps db.  Holes can include things like ponds and lakes. A future task will allow the forecaster to draw holes (or exclusion areas) in a hazard polygon. This task focuses on how holes are treated in regard to hatching and hazard clipping.
The point reduction algorithm needs to be investigated. It works well, but will occassionally produce unexpected results in areas with complex boundaries. We may want to investigate using the WarnGen point reduction algorithms if possible.
As a part of this task, I spent some time tracking down performance issues observed in hazard polygons with O(1000) points. These problems appear to be in PGEN and probably should be discussed with NCEP.
Change-Id: Ie7502fa922de22e4a5beaef8f81583bcbeac1cde