Issue #2155
Fixed bug allowing the user to deselect all choices in a radio buttons
megawidget by selecting an already-selected choice.

To test, bring up the River Flood Recommender dialog, for example, and
determine whether or not the radio buttons work as expected. When one
is clicked, it should be selected, and all others in the group should
be deselected; if the clicked radio button is already selected, nothing
should change.