Issue #2394: Transfer of RiverPro Template Functionality to Hazard Services Improved test efficiency and implemented more template variables. There are only about 8 that are not implemented now.
The RiverForecastPoints.py class provides two methods to retrieve the value(s)
corresponding to each of the legacy RiverPro template variables. One method
is named according to the template variable it emulates and one provides a
more human-readable and hopefully meaningful name. Each method is covered
by a unit test in RiverProTemplateVariableTest.java.

There is still a handful of template variables to be implemented. The
RiverProRiverFloodRecommender generates many of the values required
by the template variables when it recommends hazards. The FloodDAO is
used to access the other values which should not be in the recommender
but are required by the products (e.g. the zero datum). It is important
as this work is completed that this separation of responsibility
is kept intact. The recommender should not know about product-specific values.

See the RiverPro Template Variable Matrix for more information on these
template variables: https://docs.google.com/a/noaa.gov/spreadsheet/ccc?key=0Amz0wIe1sZZ8dHRqeHpjekxRN3VucTFMZENqbFYxYkE&usp=drive_web#gid=0

See the WHFS Support Group documentation, specifically the RiverProAppendices documentation
at https://ocwws.weather.gov/intranet/whfs/.

Change-Id: Id1e776155ee954e515dea3fd0b54abc60975ba56