Issue #2819
Fixed bugs that caused the Issue and Preview buttons in the HID to
remain disabled when they should be enabled. Then fixed merge
problems.

To test, in addition to the usual automated/unit tests, try every
conceivable mixture of proposing and issuing hazard events. This
includes:

(1) Proposing a single hazard event inside the warning area.
(2) As above, except outside the warning area.
(3) Proposing a single hazard event when another of the same type
    (e.g. FA.A) is also pending, with the first one being inside
    the warning area.
(4) As above, except outside the warning area.
(5) Repeat all of the above, except using Issue instead of Propose.

In all these cases, the various dialogs that come up as a result
should be tested to ensure that regardless of whether the user
cancels the dialog (via either the titlebar X button or via the
Dismiss or Cancel buttons), or continues (i.e. Issue within the
Product Editor, Continue within the Product Staging Dialog, Yes
within the Issue Confirmation messagebox dialog), the Preview and
Issue buttons are always re-enabled when they should be.

Change-Id: I560dac86e3731b0c15141711b6ee589b2fa78817