Issue #1633 Product Parts and Simplified Product Dictionary -- initial installment
NOTE: In this code review, some products have editable fields and some do not.  This seemed to cause problems for the Product Editor dialog. Throughthe debugger, it appears the Product Editor is indexing the tabs and may be keeping parallel arrays for storing the edited information.  We got an out of range index occurring at ProductGenerationDialog line 652 during the course of running the Mixed Hazard Story and trying to edit the cityList.  
Thus, we are temporarily bypassing editable functionality in the Product Generators to allow Jonathan to further test the Product Editor.  To re-enable, remove bypass lines in Legacy_ProductGenerator method setEditedField.

Change-Id: I2df03a06e9254a3a3887206a5d46deaacbbcf81a