Issue #2925
First phase of implementing class-based metadata changes. Almost
everything that needs to be done for the latter has been done as part
of this changeset, with the exception of reimplementing the HID to
actually use the dynamic metadata as it changes in concert with state
and type changes to hazard events. Changes included in this changeset:

Session manager modified to load class-based metadata when a hazard
event's type or state changes.

Session events changed to carry more information about what
modifications are made to hazard events.

Much of the business logic found in the HID package, the message
handler, etc. has been moved to the session manager where it belongs.

Changed HID manipulation of the session manager to be direct rather
than via HazardDetailAction messages in most cases.

Large-scale changes to the megawidget framework, specifically to allow
megawidget specifiers to always provide default starting state values
when the specifications do not include these, and to allow the merging
of starting state values from specifiers with an existing set of state
values.

Amended to include appropriate Python path in configuration manager
for Jep creation.

Change-Id: I262e6f93dc6bf2dedc8cf42a7122f303dbf700af