Issue #2925
Major refactor of the Hazard Detail View (or HID) and supporting
classes:

The view part that the user actually sees, its view delegate, and that
view's presenter have been almost totally rewritten to support dynamic
metadata, that is, metadata widgets that get reloaded each time a
hazard event's state changes (though they may be easily changed to
reload in response to other notifications in the future if need be).

Since this required such a major refactor, the view part, view and
presenter were all changed to use (with a couple of small exceptions)
the direct-model-manipulation strategy. The view and the presenter are
very loosely coupled, which was not the case before; and the view knows
nothing about the model (session manager). The presenter changes the
model directly, instead of via messages, and the model notifies the
presenter of changes to the model's state via the event bus. The HID
should now be ready for the switchover to use of a worker thread when we
get to that point, since the view is careful to move between a worker
thread and the UI thread as necessary. (Currently, as with the other
parts of the GUI, it simply uses the UI thread for everything.)

Finally, some changes in the view part itself have been made, including
the disabling of the Propose button for anything that has been issued or
ended; the rendering of the category selector read-only if the event has
been issued or ended; the extraction of the hazard type and time range
widgets from the scrollable area of the dialog, the shortening of labels
for those widgets so that they can fit on a narrow view part, and the
addition of a horizontal scrollbar for the metadata section when
necessary.

Added code to fix two bugs, both intermittent:

(1) Complaints about a view part not being created.
(2) Problems with SessionConfigurationManager Jep instance closing
    unexpectedly.

Also made minor changes in response to Matt's code review.

Added code to fix problem of exceptions being thrown when changing
perspectives.

Added cleanup of Jep instance before latter is closed.

Added Dan's fixes for the automated tests. The latter will remain
relatively brittle, in my opinion, until we have the time to really
characterize the notification propagation that occurs in response to
various changes initiated by the user (e.g. change of hazard end
time, or ending a hazard, or whatever).

Change-Id: I7937af0710e61dabb4eefa6fde0eead22dfceeea