Issue #2155
Changed spinner megawidgets that include scale widgets so that the
latter have their handles snap to the current value when the user
drags them and releases the mouse. Previously, they did not snap to
the current value, which could look confusing.

Fixed a bug in the time delta megawidget that caused it to interpret
the starting value as being specified in milliseconds, instead of the
unit value provided in that megawidget's specification.

Fixed bug in time scale megawidget that caused it to treat the
timeDescriptors parameter as a mandatory part of its specification.

Fixed minor problem with invalid incrementDelta parameter error message
in the fractional spinner megawidget. Also changed to use correct
bounds-checking for incrementDelta (previously, it was determining the
bounds incorrectly, since it was doing this before checking the
precision).

Added missing "choices" option to set of mutable properties for
hierarchical tree choice megawidget.

Fixed bug in stateful megawiget specifier that caused attempts to use
default values for metadata megawidgets to silently fail.

Added detailed test instructions to the corresponding Redmine issue.

Change-Id: I31e1fb98d41987594f008724516bfc1adfc29698