Issue #4010
Added the missing "expand bar" megawidget, which is a container of other
megawidgets that allows its pages (within which the child megawidgets
are embedded) to have their visibility toggled by clicking on the header
bars for each of the pages. Essentially this is a wrapper for the SWT
ExpandBar.

Since the expand bar megawidget introduced a new type of megawidget that
can change its size during its lifetime, code was added to allow
megawidget managers and their clients to listen for Composite resizes
and respond accordingly. For example, the metadata panel in the HID
responds to any such resizes by updating its scrolled composite so that
the latter knows the new size of its client area.

Altered the parameters editor factory to provide the option of embedding
different parameter editors (megawidgets) within expand bars if desired.

Changed the megawidget editor demo test application to include a
scrollable area and to use expand-bar-embedded megawidgets in order to
allow this new functionality to be tested.

Took opportunity to remove the extra callback parameter from notifier
megawidgets, which meant that simplification of the code base could be
done (removal of the INotifierSpecifier and NotifierMegawidgetSpecifier,
and anything related to the extra callback parameter). This was done
because the removed code had been rendered useless by earlier changes to
the framework.

Appended: Somehow Eclipse didn't show that one file no longer compiled,
so I only discovered this after the push of the changeset. This version
fixes that compilation problem.

Change-Id: Iee610f47d4d46b8e9865755c08ff345285af4f74