Issue #4009
Added ability to store and retrieve arbitrary "extra data" within each
megawidget. This may then be used by interdependency scripts to persist
arbitrary, non-megawidget-specific information between script
executions.

Used new capability to implement the "fallBelow" functionality with
regard to the user toggling "Until further notice" (UFN) on or off. The
script used to merely change the "fallBelow" slider to be read-only;
now, it also sets the time to the special "UFN" value, or back to what
it was before "UFN" was last turned on. It stores the old value as an
arbitrary piece of data attached to the "riseAbove:crest:fallBelow"
megawidget.

Removed functionality corresponding to what was implemented in the
script from the session event manager. This gets rid of special-case
code that really did not belong in that class, or in any hard-coded Java
class.