Issue #4184
Fixed argument specification for megawidget test launch configurations
to work with 14.4 (it was pointing at 14.3 location for a library that
is needed to launch properly).

Also fixed problem with old SWT library being included in the tests
lib folder, which made some of the megawidget standalone tests fail.

To test, follow step 1 of the test steps listed in the Redmine
description of Defect #3982, simply to ensure that the launch
configuration for the megawidget tests is now correct. Also run
Hazard Services and ensure that the HID comes up normally (since
any failure would be visible and catastrophic, i.e. no HID would come
up, or it would come up with missing widgets).

Change-Id: I006a7b3262dbca80c17d5bdd755e1c26f4fa28f0