Issue #3512
Added a new time range megawidget. This is similar to the time scale
megawidget, with the principal differences being (a) it must always have
two states associated with it, a lower time boundary and an upper time
boundary; and (b) it has a drop-down selector holding the time deltas
for choosing the upper time (i.e. the upper boundary is the lower
boundary plus the time delta).

If the state is set programmatically (or via the specifier when the
megawidget is created) to have an upper boundary with a time delta from
the lower boundary that is not one of the selectable time deltas in the
drop-down selector, the latter reads "N/A", and the two times may be
modified separately. If the state changes so that the time delta is now
one of the selectable ones, the megawidget has its scale widget lock its
interval so that when the upper boundary thumb is dragged, the lower one
moves with it in concert, and vice versa.

The megawidget supports detail fields just like the time scale. It has
similar mutable properties, with the addition of the list of selectable
time deltas (so that they may be changed during the lifetime of the
megawidget).

Changes were required to the megawidget framework to allow multiple
simultaneous state changes to occur, since both boundaries may change in
concert. This meant fairly widespread changes (in terms of number of
files touched). The megawidget manager, which was affected by this
significantly, was therefore simplified to take a listener when created
to be notified of state changes, etc. instead of having the manager be
an abstract class that had to be subclassed whenever used. This is a
design correction; my abstract class scheme was proving to be
cumbersome.

Test steps for this changeset are listed in the Redmine issue. Note that
the time range megawidget is not actually used by Hazard Services yet;
the next changeset for this issue will place it within the Hazard
Information Dialog when appropriate for a particular hazard type.

Note that this code review is a simple translation of the now-abandoned
code review at https://collaborate3.nws.noaa.gov/code-review/#/c/1100/
to the new 14.4.x directory structure. All changes through patch set 7
from the earlier one are present here, including fixes in response to
Dan's concerns.