Issue #3512
(Second attempt at this code review; the first was merged and then
dropped, so this is being submitted as a new changeset, but the comments
made in https://collaborate3.nws.noaa.gov/code-review/#/c/1163 apply to
this code review as well.)

Changed the HID to use a duration drop-down selector instead of an
absolute end time selector for hazard types that specify a list of
possible durations within the HazardTypes.py file. Where appropriate,
such hazard types also support the setting of end time to "Until Further
Notice", just as with absolute end times.

Since a default duration is needed that was hazard-type-specific so that
when a hazard is created, it starts off with a duration that is one of
the duration choices listed for that type in HazardTypes.py, the default
duration setting within the latter file is now used as the starting
duration.

Tests are as listed in the latest comments (notes) for the Redmine issue.

Changed to restore the static import lines as per Hazard Services canon.

Change-Id: I79a4f5b5ae223db3aaba816634e4e64308fff1b0