Issue #3512
Changed the HID to use a duration drop-down selector instead of an
absolute end time selector for hazard types that specify a list of
possible durations within the HazardTypes.py file. Where appropriate,
such hazard types also support the setting of end time to "Until Further
Notice", just as with absolute end times.

Since a default duration is needed that was hazard-type-specific so that
when a hazard is created, it starts off with a duration that is one of
the duration choices listed for that type in HazardTypes.py, the default
duration setting within the latter file is now used as the starting
duration.

Tests are as listed in the latest comments (nots) for the Redmine issue.

Change-Id: I520ac8122641709f23cce87b17e3e0f3947304ff