Issue #4259
Fixed problems with spinner megawidgets and negative values (spinner
text field not wide enough to show value with minus sign in front of it,
and unable to hold negative values if a scale widget was included in the
megawidget's configuration).

Also removed use of the notification logger from the automated tests
class, since it was dramatically slowing down performance when for
example dragging an event end time back and forth in the HID (the
console's corresponding end time thumb would lag way behind).

Changeed to be based on new #3512 code review.

Change-Id: If0b017b19fde6b7df41288c33040b24db049edf9