Issue #4098
Added new SWT wrapper megawidget. This is a special megawidget that is
useless to focal points, but which may have arbitrary SWT widgets added
to it from within Java code.

An example usage of this megawidget is found within SwtWrapper.json
(giving the JSON used to specify its properties), and at the bottom of
the MegawidgetTest.createDialogArea() method, within the "if
(swtWrapperIdentifier != null)" block. A button is created as a child
of the wrapper's Composite.

See the Redmine issue for information on how to test this changeset.