Issue #3525.  Automated tests for product generators. Be sure to add the following line to /awips2/eclipse/eclipse.sh:
export HAZARD_SERVICES_SOURCE=$ADE/hazardServices

This assumes you are following our convention of defining $ADE as the
directory into which we download A2 and hazards source code.

This is a start but plenty of work to do.
It does not handle product generation involving multiple hazards.
It needs to be able to accept commands to clear the active table.
Right now it clears the table between every life cycle test.
It needs more tests filled in of course.
In fact, the flw_fls test is merely a copy of the ffw_ffs.
If the configuration for the suite of tests contains no entries
then it needs to run all the tests.
It needs to handle product dialog information.
Change-Id: I24666f0445553e544d8668f7e1d0b59059178025