Issue #4560 Event Metadata Reload upon Metadata State Change
Added ability to trigger a full metadata reload for a hazard event via
changes to metadata attribute values. The values are indicated by
including the parameter "refreshMetadata" in the specifier for any
megawidget created to manipulate metadata values for a hazard event in
the HID. If such a megawidget has its state changed, the metadata is
reloaded.

Change-Id: I737616f0d908c816df4dd447f8bfd91a5a9a37bf