Issue #3627 - Reposting Dan's changes with other comments addressed.
Correction to product generation optimization.

Also, cleanup/refactoring.  In particular, eliminated the
HazardServicesProductGenerationHandler class and moved its functionality
into SessionProductManager.  Also, reorganized SessionProductManager.

Now realized with help from Chris that I was not improperly using the cache
in SessionProductManager#createProductsFromProductStagingInfo so I
changed the code so that it does not make that correction after all.