Issue #5050 Hidden field megawidget
Added a hidden field megawidget. This can be used much as a hidden input
field is used in HTML documents: It can hold state that is not exposed
to the user via the GUI. Furthermore, it can have its state changed via
interdependency scripts as a result of other megawidget state changes.
The changeset includes a test case with an interdependency script,
HiddenFieldScript.py, that may be used as a guide as to how to do this.

Also cleaned up a few very minor things in the megawidgets code.

Change-Id: I929ad5881fbd0cf82244d13af0b7f3328da31a62