Issue #4042 Reimplementation of Product Staging Dialog
The Product Staging Dialog is now a two-step dialog. It will now, when a
preview or issue is attempted, show zero, one, or both of the following
steps: (a) display a checklist of hazard events that may be included
with each product being issued, if there are unselected events that
could be included; and (b) display product-generator-specific
megawidgets to allow staging information to be input by the user. The
second stage includes scrollbars when needed, and allows interdependency
scripts to be used with the megawidgets the product generator chooses to
display.

Refactored the session product manager to support this up-to-two-step
process, and simplified its interface to make fewer of its methods
public.

Fixed a bug that caused generation to continue to be attempted, at the
very least wasting CPU cycles (and perhaps doing worse things than
that), when an event was found to be outside the warning area and
product generation had supposedly halted.

Uncommented code in the FFA and FLW_FLS product generators to allow them
to show megawidgets for the step two of the Product Staging Dialog. Also
wrote interdependency script to allow these generators to have their
megawidgets interact with one another (use of menu button megawidget
allows the filling in of canned synopsis text in a text area which can
then be edited).

Added new option to Label megawidget to allow it to have a preferred
width (given as characters) when laid out.

Added new MenuButton megawidget, allowing the creation of a button with
a dropdown menu, which in turn includes one or more commands that may be
issued. No state is tracked, unlike the ComboBox.

Made changes in response to Mike Duff's comments. Also fixed a couple of
bugs: (1) Attempting to generate a product with at least one hazard
event selected with no hazard type resulted in an exception; and (2)
Deselecting all hazard events in step 1 of the Product Staging Dialog
would leave the HID's Preview and Issue buttons disabled, even though no
product generation was to be done in that case.

Commented out product-specific megawidget specification in the
defineDialog() method for every product generator that was specifying
such. These need to be revisited in a subsequent code review once the
appropriate megawidgets for each product generator are determined.

Fixed bug introduced by previous patchset in session product manager
causing no product generation to occur if only step 1 of the product
staging dialog was needed.

Fixed bug that manifested itself only on machines with Python debugging
enabled in which the lack of the hazardMetaData directory on the product
generator Python include path could cause product generation to fail.

Change-Id: I2876fe52543020f7bb2e053f52e07b0bba9d1ef5